analytics-service/requirements.txt
grpcio==1.63.0
grpcio-tools==1.63.0
grpcio-reflection==1.63.0
protobuf==5.26.1
googleapis-common-protos==1.63.0
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-grpc==0.45b0
pyjwt==2.8.0
structlog==24.1.0

# Test & lint
pytest==8.2.0
pytest-cov==5.0.0
ruff==0.4.4
freezegun==1.5.1
